# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EVE Online PAP Tracking System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /characters/{id}:
    delete:
      operationId: deleteCharacter
      tags: [characters]
      summary: Unlink a character and retain site data for up to 3 months
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Character unlinked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteCharacterResponse"
        "400":
          description: Invalid request
        "403":
          description: Forbidden
  /admin/users/{id}/characters:
    get:
      operationId: getUserCharacters
      tags: [characters]
      summary: Get all characters for a specific user (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of characters for the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Character"
...
  /battle-reports/{id}:
    get:
      operationId: getBattleReport
      tags: [battle-reports]
      summary: Get a fleet battle report (all authenticated members)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Battle report with fleet and killmail details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BattleReportDetail"
        "404":
          description: Battle report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /battle-reports/{id}/refresh:
    post:
      operationId: refreshBattleReport
      tags: [battle-reports]
      summary: Re-sync a fleet battle report (FC and above)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "202":
          description: Battle report refresh queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                required:
                  - status
  /command/battle-replays:
    get:
      operationId: listBattleReplays
      tags: [battle-replays]
      summary: List battle reports with FC review status (FC and above)
      responses:
        "200":
          description: Battle reports available for command review
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BattleReplaySummary"
        "403":
          description: FC role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /command/battle-replays/{id}:
    get:
      operationId: getBattleReplay
      tags: [battle-replays]
      summary: Get the complete FC battle-review workbench (FC and above)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Battle report and private review draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BattleReplayDetail"
        "404":
          description: Battle report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateBattleReplay
      tags: [battle-replays]
      summary: Save or publish FC nodes, annotations, and conclusion (FC and above)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBattleReplayBody"
      responses:
        "200":
          description: Saved battle review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BattleReview"
        "400":
          description: Invalid review content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /command/battle-replays/{id}/analyze:
    post:
      operationId: analyzeBattleReplay
      tags: [battle-replays]
      summary: Run evidence-based AI battle analysis (FC and above)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "202":
          description: Analysis queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [generating]
                required:
                  - status
        "404":
          description: Battle report not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
        - error
...
    Character:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        eveCharacterId:
          type: integer
        eveCharacterName:
          type: string
        corporationId:
          type: ["integer", "null"]
        corporationName:
          type: ["string", "null"]
        isMain:
          type: boolean
        deletedAt:
          type: ["string", "null"]
          format: date-time
        retainedUntil:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - eveCharacterId
        - eveCharacterName
        - isMain
        - createdAt
    DeleteCharacterResponse:
      type: object
      properties:
        success:
          type: boolean
        removedMain:
          type: boolean
        newMainCharacterId:
          type: ["integer", "null"]
        newMainCharacterName:
          type: ["string", "null"]
        requiresReauthentication:
          type: boolean
      required:
        - success
        - removedMain
        - newMainCharacterId
        - newMainCharacterName
        - requiresReauthentication
...
    BattleReportSummary:
      type: object
      properties:
        id:
          type: integer
        fleetId:
          type: ["integer", "null"]
        fleetName:
          type: string
        fleetCommander:
          type: string
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, generating, ready, partial, failed]
        errorMessage:
          type: ["string", "null"]
        totalDestroyed:
          type: number
        totalLost:
          type: number
        damageDealt:
          type: number
        killmailCount:
          type: integer
        friendlyLosses:
          type: integer
        hostileLosses:
          type: integer
        primarySystemId:
          type: ["integer", "null"]
        primarySystemName:
          type: ["string", "null"]
        generatedAt:
          type: ["string", "null"]
          format: date-time
        lastSyncedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        participantCount:
          type: integer
        systemCount:
          type: integer
      required:
        - id
        - fleetName
        - fleetCommander
        - startedAt
        - endedAt
        - status
        - totalDestroyed
        - totalLost
        - damageDealt
        - killmailCount
        - friendlyLosses
        - hostileLosses
        - createdAt
        - participantCount
        - systemCount
    BattleReportParticipant:
      type: object
      properties:
        id:
          type: integer
        eveCharacterId:
          type: integer
        characterName:
          type: string
        corporationId:
          type: ["integer", "null"]
        corporationName:
          type: ["string", "null"]
        primaryShipTypeId:
          type: ["integer", "null"]
        primaryShipName:
          type: ["string", "null"]
        damageDealt:
          type: number
        killsInvolved:
          type: integer
        finalBlows:
          type: integer
        losses:
          type: integer
      required:
        - id
        - eveCharacterId
        - characterName
        - damageDealt
        - killsInvolved
        - finalBlows
        - losses
    BattleReportSystem:
      type: object
      properties:
        solarSystemId:
          type: integer
        solarSystemName:
          type: ["string", "null"]
        killmailCount:
          type: integer
        friendlyLosses:
          type: integer
        hostileLosses:
          type: integer
        destroyedValue:
          type: number
        lostValue:
          type: number
      required:
        - solarSystemId
        - killmailCount
        - friendlyLosses
        - hostileLosses
        - destroyedValue
        - lostValue
    BattleReportAttacker:
      type: object
      properties:
        characterId:
          type: ["integer", "null"]
        characterName:
          type: ["string", "null"]
        corporationId:
          type: ["integer", "null"]
        corporationName:
          type: ["string", "null"]
        allianceId:
          type: ["integer", "null"]
        allianceName:
          type: ["string", "null"]
        shipTypeId:
          type: ["integer", "null"]
        shipName:
          type: ["string", "null"]
        damageDone:
          type: integer
        finalBlow:
          type: boolean
        isFleetMember:
          type: boolean
      required:
        - damageDone
        - finalBlow
        - isFleetMember
    BattleReportKillmail:
      type: object
      properties:
        id:
          type: integer
        killmailId:
          type: integer
        killmailTime:
          type: string
          format: date-time
        solarSystemId:
          type: integer
        solarSystemName:
          type: ["string", "null"]
        victimCharacterId:
          type: ["integer", "null"]
        victimCharacterName:
          type: ["string", "null"]
        victimCorporationId:
          type: ["integer", "null"]
        victimCorporationName:
          type: ["string", "null"]
        victimAllianceId:
          type: ["integer", "null"]
        victimAllianceName:
          type: ["string", "null"]
        victimShipTypeId:
          type: integer
        victimShipName:
          type: ["string", "null"]
        victimIsFleetMember:
          type: boolean
        totalValue:
          type: number
        damageTaken:
          type: integer
        friendlyDamage:
          type: integer
        friendlyAttackers:
          type: integer
        finalBlowByFleet:
          type: boolean
        attackers:
          type: array
          items:
            $ref: "#/components/schemas/BattleReportAttacker"
        zkillboardUrl:
          type: ["string", "null"]
      required:
        - id
        - killmailId
        - killmailTime
        - solarSystemId
        - victimShipTypeId
        - victimIsFleetMember
        - totalValue
        - damageTaken
        - friendlyDamage
        - friendlyAttackers
        - finalBlowByFleet
        - attackers
    BattleReportDetail:
      allOf:
        - $ref: "#/components/schemas/BattleReportSummary"
        - type: object
          properties:
            participants:
              type: array
              items:
                $ref: "#/components/schemas/BattleReportParticipant"
            killmails:
              type: array
              items:
                $ref: "#/components/schemas/BattleReportKillmail"
            systems:
              type: array
              items:
                $ref: "#/components/schemas/BattleReportSystem"
            publishedReview:
              oneOf:
                - $ref: "#/components/schemas/BattleReview"
                - type: "null"
          required:
            - participants
            - killmails
            - systems
    BattleReviewManualNode:
      type: object
      properties:
        id:
          type: string
          maxLength: 100
        category:
          type: string
          enum: [key_ship, key_kill, loss_peak, phase, note]
        occurredAt:
          type: string
          format: date-time
        title:
          type: string
          maxLength: 160
        description:
          type: string
          maxLength: 2000
      required: [id, category, occurredAt, title, description]
    BattleReplayKeyEvent:
      type: object
      properties:
        killmailId:
          type: integer
        occurredAt:
          type: string
          format: date-time
        title:
          type: string
        reason:
          type: string
        confidence:
          type: number
        shipName:
          type: ["string", "null"]
        pilotName:
          type: ["string", "null"]
        friendlyLoss:
          type: boolean
        totalValue:
          type: number
        evidenceLevel:
          type: string
          enum: [confirmed, inferred]
        evidence:
          type: string
      required:
        [
          killmailId,
          occurredAt,
          title,
          reason,
          confidence,
          friendlyLoss,
          totalValue,
        ]
    BattleReplayLossPeak:
      type: object
      properties:
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
        title:
          type: string
        reason:
          type: string
        confidence:
          type: number
        killmailIds:
          type: array
          items:
            type: integer
        friendlyLosses:
          type: integer
        hostileLosses:
          type: integer
        totalValue:
          type: number
        evidenceLevel:
          type: string
          enum: [confirmed, inferred]
        evidence:
          type: string
      required:
        [
          startedAt,
          endedAt,
          title,
          reason,
          confidence,
          killmailIds,
          friendlyLosses,
          hostileLosses,
          totalValue,
        ]
...
    BattleReplayAnalysis:
      type: object
      properties:
        version:
          type: integer
          enum: [1]
        source:
          type: string
          enum: [openai, rules]
        model:
          type: string
        generatedAt:
          type: string
          format: date-time
        summary:
          type: string
        keyShips:
          type: array
          items:
            $ref: "#/components/schemas/BattleReplayKeyEvent"
        keyKills:
          type: array
          items:
            $ref: "#/components/schemas/BattleReplayKeyEvent"
        lossPeaks:
          type: array
          items:
            $ref: "#/components/schemas/BattleReplayLossPeak"
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/BattleReplaySuggestion"
        phases:
          type: array
          items:
            $ref: "#/components/schemas/BattleReplayPhase"
        dataQuality:
          $ref: "#/components/schemas/BattleReplayDataQuality"
      required:
        [
          version,
          source,
          model,
          generatedAt,
          summary,
          keyShips,
          keyKills,
          lossPeaks,
          suggestions,
        ]
    BattleReview:
      type: object
      properties:
        status:
          type: string
          enum: [draft, published]
        aiStatus:
          type: string
          enum: [not_started, generating, ready, failed]
        aiSource:
          type: ["string", "null"]
          enum: [openai, rules, null]
        aiModel:
          type: ["string", "null"]
        aiError:
          type: ["string", "null"]
        aiAnalysis:
          oneOf:
            - $ref: "#/components/schemas/BattleReplayAnalysis"
            - type: "null"
        manualNodes:
          type: array
          items:
            $ref: "#/components/schemas/BattleReviewManualNode"
        conclusion:
          type: string
        aiAnalyzedAt:
          type: ["string", "null"]
          format: date-time
        publishedAt:
          type: ["string", "null"]
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [status, aiStatus, manualNodes, conclusion, updatedAt]
    BattleReplaySummary:
      allOf:
        - $ref: "#/components/schemas/BattleReportSummary"
        - type: object
          properties:
            reviewStatus:
              type: string
              enum: [not_started, draft, published]
            aiStatus:
              type: string
              enum: [not_started, generating, ready, failed]
            reviewUpdatedAt:
              type: ["string", "null"]
              format: date-time
          required: [reviewStatus, aiStatus]
    BattleReplayDetail:
      allOf:
        - $ref: "#/components/schemas/BattleReportDetail"
        - type: object
          properties:
            review:
              oneOf:
                - $ref: "#/components/schemas/BattleReview"
                - type: "null"
          required: [review]
    UpdateBattleReplayBody:
      type: object
      properties:
        status:
          type: string
          enum: [draft, published]
        conclusion:
          type: string
          maxLength: 10000
        manualNodes:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/BattleReviewManualNode"
      required: [status, conclusion, manualNodes]
...